chore: rename theme.domain
