+++ frontend/core/hooks/useTheme.ts
@@ -1,6 +1,6 @@
 import THEME, { LOCAL_THEME_KEY, THEME_MODE } from '~/const/theme'
 import type { TThemeMode, TThemeName } from '~/spec'
-import useThemeDomain from '~/stores/theme.domain/hooks'
+import useThemeDomain from '~/stores/theme/hooks'
 
 type TRet = {
   theme: TThemeName
