--- frontend/core/stores/theme.domain/spec.d.ts
@@ -1,11 +1,0 @@
-import type { TThemeMode, TThemeName } from '~/spec'
-
-export type TStore = {
-  theme: TThemeName
-  themeMode: TThemeMode
-  // actions
-  change: (theme: TThemeName) => void
-  changeMode: (theme: TThemeMode) => void
-}
-
-export type TInit = TThemeName
